se realizan pruebas finales

+++ app/src/main/java/com/psilva/comunicafacil/ui/screens/HomeScreen.kt
@@ -88,7 +88,7 @@ fun HomeScreen(onCerrarSesion: () -> Unit) {
                 tipoMensaje = TipoMensaje.INFO
                 estadoSnackbar.showSnackbar(
                     message = "TTS no disponible en este dispositivo/emulador",
-                    duration = SnackbarDuration.Long
+                    duration = SnackbarDuration.Short
                 )
 
             }
